Generify menu system

+++ GameLib/src/main/kotlin/fr/o80/gamelib/loop/MouseButtonPipeline.kt
@@ -1,5 +1,5 @@
 package fr.o80.gamelib.loop
 
 interface MouseButtonPipeline {
-    fun onButton(button: Int, action: Int, block: (x: Float, y: Float) -> Unit)
+    fun onButton(button: Int, action: Int, block: (x: Double, y: Double) -> Unit)
 }
