+++ GameLib/src/main/kotlin/fr/o80/gamelib/loop/MouseButtonPipelineImpl.kt
@@ -4,21 +4,21 @@ import org.lwjgl.glfw.GLFWMouseButtonCallbackI
 
 internal class MouseButtonPipelineImpl : MouseButtonPipeline, GLFWMouseButtonCallbackI {
 
-    private val callbacks = mutableListOf<Triple<Int, Int, (Float, Float) -> Unit>>()
+    private val callbacks = mutableListOf<Triple<Int, Int, (Double, Double) -> Unit>>()
 
-    private var mouseX: Float = -1f
-    private var mouseY: Float = -1f
+    private var mouseX: Double = -1.0
+    private var mouseY: Double = -1.0
 
     override fun invoke(window: Long, button: Int, action: Int, mods: Int) {
         callbacks.filter { (k, a, _) -> k == button && a == action }
             .forEach { it.third(mouseX, mouseY) }
     }
 
-    override fun onButton(button: Int, action: Int, block: (x: Float, y: Float) -> Unit) {
+    override fun onButton(button: Int, action: Int, block: (x: Double, y: Double) -> Unit) {
         callbacks.add(Triple(button, action, block))
     }
 
-    fun onMouseMove(x: Float, y: Float) {
+    fun onMouseMove(x: Double, y: Double) {
         mouseX = x
         mouseY = y
     }
